wasm3: implement mandatory WebAssembly Core 3.0 (#278)

+++ src/wago/funcref_global_test.go
@@ -249,8 +249,8 @@ func TestInstantiateRejectsUnsupportedReferenceGlobalMetadata(t *testing.T) {
 		{name: "non-structural initializer bits", c: &Compiled{Globals: []GlobalDef{{Type: ValFuncRef, Bits: 1}}}, want: "non-structural funcref global initializer"},
 		{name: "non-null externref initializer", c: &Compiled{Globals: []GlobalDef{{Type: ValExternRef, Bits: 1}}}, want: "non-null externref global initializer"},
 		{name: "ref.func wrong type", c: &Compiled{Globals: []GlobalDef{{Type: ValI64, HasInitFunc: true}}, NeedsFuncRefDescs: true}, want: "ref.func initializer has type i64"},
-		{name: "ref.func out of range", c: &Compiled{Code: []byte{0}, Entry: []int{0}, Funcs: []FuncSig{{}}, FuncTypeID: []uint32{0}, Globals: []GlobalDef{{Type: ValFuncRef, HasInitFunc: true, InitFunc: 1}}, NeedsFuncRefDescs: true}, want: "ref.func initializer index 1 out of range"},
-		{name: "ref.func without descriptors", c: &Compiled{Code: []byte{0}, Entry: []int{0}, Funcs: []FuncSig{{}}, FuncTypeID: []uint32{0}, Globals: []GlobalDef{{Type: ValFuncRef, HasInitFunc: true}}}, want: "ref.func initializer without descriptor arena"},
+		{name: "ref.func out of range", c: &Compiled{Code: []byte{0}, Entry: []int{0}, Funcs: []FuncSig{{}}, FuncTypeID: []uint64{0}, Globals: []GlobalDef{{Type: ValFuncRef, HasInitFunc: true, InitFunc: 1}}, NeedsFuncRefDescs: true}, want: "ref.func initializer index 1 out of range"},
+		{name: "ref.func without descriptors", c: &Compiled{Code: []byte{0}, Entry: []int{0}, Funcs: []FuncSig{{}}, FuncTypeID: []uint64{0}, Globals: []GlobalDef{{Type: ValFuncRef, HasInitFunc: true}}}, want: "ref.func initializer without descriptor arena"},
 	}
 	for _, tc := range tests {
 		t.Run(tc.name, func(t *testing.T) {
